feat: add stack to project with kind
This commit will add a stack to the project. Examples of stack is Laravel, React, etc. Basically anything I use to create a project is what I consider a stack. I also have a kind model which just describes what type software a stack is to me. For example Laravel is a fullstack framework so the kind will fall under Fullstack.