feat: add links to projects with link type
I added a lnks model that allows me to add external links to my projects. This allows e to have links to the github of the project, links to a small video presentation of the project or even links to the project live site. I also added a link type to help me easily organize my links but this is only an internal implementation for me. Examples of link types are Github, Youtube, Live Site, etc.